requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
grpcio==1.65.1
protobuf==5.27.2
